airbyte-ci: do not eagerly fetch connector secrets
